+++ extensions/Worker.Extensions.CosmosDB/src/Config/CosmosDBBindingOptions.cs
@@ -9,6 +9,12 @@ using Microsoft.Azure.Functions.Worker.Extensions.CosmosDB;
 
 namespace Microsoft.Azure.Functions.Worker
 {
+     /// <summary>
+    /// Internal options for configuring the CosmosDB binding.
+    /// This class is used internally by the Azure Functions runtime to manage the CosmosDB connection and clients.
+    /// It is not intended to be used directly in user code.
+    /// Any public configuration options should be set on the <see cref="CosmosDBOptions"/> class, which is publicly accessible.
+    /// </summary>
     internal class CosmosDBBindingOptions
     {
         public string? ConnectionName  { get; set; }
@@ -21,7 +27,7 @@ namespace Microsoft.Azure.Functions.Worker
 
         public CosmosSerializer? Serializer { get; set; }
 
-        public CosmosClientOptions? CosmosClientOptions { get; set; }
+        public CosmosDBOptions? CosmosDBOptions { get; set; }
 
         internal string BuildCacheKey(string connection, string region) => $"{connection}|{region}";
 
@@ -34,18 +40,23 @@ namespace Microsoft.Azure.Functions.Worker
                 throw new ArgumentNullException(nameof(ConnectionName));
             }
 
+            if (CosmosDBOptions is null)
+            {
+                CosmosDBOptions = new CosmosDBOptions();
+            }
+
             string cacheKey = BuildCacheKey(ConnectionName!, preferredLocations);
 
             // Do not override if preferred locations is configured via CosmosClientOptions
-            if (!string.IsNullOrEmpty(preferredLocations) && CosmosClientOptions.ApplicationPreferredRegions is null)
+            if (!string.IsNullOrEmpty(preferredLocations) && CosmosDBOptions.ClientOptions.ApplicationPreferredRegions is null)
             {
-                CosmosClientOptions.ApplicationPreferredRegions = Utilities.ParsePreferredLocations(preferredLocations);
+                CosmosDBOptions.ClientOptions.ApplicationPreferredRegions = Utilities.ParsePreferredLocations(preferredLocations);
             }
 
             // Do not override if the serializer is configured via CosmosClientOptions
-            if (Serializer is not null && CosmosClientOptions.Serializer is null)
+            if (Serializer is not null && CosmosDBOptions.ClientOptions.Serializer is null)
             {
-                CosmosClientOptions.Serializer = Serializer;
+                CosmosDBOptions.ClientOptions.Serializer = Serializer;
             }
 
             return ClientCache.GetOrAdd(cacheKey, (c) => CreateService());
@@ -54,8 +65,8 @@ namespace Microsoft.Azure.Functions.Worker
         private CosmosClient CreateService()
         {
             return string.IsNullOrEmpty(ConnectionString)
-                    ? new CosmosClient(AccountEndpoint, Credential, CosmosClientOptions) // AAD auth
-                    : new CosmosClient(ConnectionString, CosmosClientOptions); // Connection string based auth
+                    ? new CosmosClient(AccountEndpoint, Credential, CosmosDBOptions?.ClientOptions) // AAD auth
+                    : new CosmosClient(ConnectionString, CosmosDBOptions?.ClientOptions); // Connection string based auth
         }
     }
 }